feat: add postgress dockerfile to dataBase that was moved on src level + refactor: network in docker-compose file correction
