+++ .env
@@ -1,10 +1,21 @@
 NODE_ENV=development
 PORT=4000
+
 # Set LOG LEVEL to: debug | error | info | wanr | fatal | trace
 LOG_LEVEL=trace
+
 # Set Docker Project name
 COMPOSE_PROJECT_NAME=trello-clone
+
+# Set postgress details
+POSTGRES_USER=dbuser
+POSTGRES_PASSWORD=admin2021
+POSTGRES_DB=trello-clone
+POSTGRES_PORT=5432
+
+# Set authorization details
 AUTH_MODE=false
 JWT_SECRET_KEY=secret-key
+
 # Set your database connection information here
 MONGO_CONNECTION_STRING=your-mongo-db-connection-string
