+++ dataBase/docker/Dockerfile
@@ -1,0 +1,1 @@
+FROM postgres:13.5-alpine
