+++ nodemon.json
@@ -1,0 +1,9 @@
+{
+    "verbose": true,
+    "ignore": ["test/*", "docs/*"],
+    "watch": ["src/**/*.ts"],
+    "legacyWatch": true,
+    "execMap": {
+      "ts": "node --inspect=0.0.0.0:9229 --nolazy -r ts-node/register"
+    }
+  }
