+++ package.json
@@ -11,6 +11,8 @@
     "start": "node dist/server.js",
     "build": "tsc",
     "start:fast": "npm run build && npm run start",
+    "start:watch": "tsc-watch --project . --outdir ./dist --onSuccess \" nodemon ./dist/server.js\"",
+    "dev": "nodemon -L src/server.ts",
     "lint": "eslint --ext .js,.ts --fix src/",
     "test": "cross-env DEBUG=rs:* jest --testMatch \"<rootDir>/test/e2e/test/*.test.js\" --noStackTrace --runInBand",
     "test:auth": "cross-env DEBUG=rs:* TEST_MODE=auth jest --noStackTrace",
@@ -53,6 +55,8 @@
     "fastify-swagger": "^4.12.6",
     "pino": "^7.6.0",
     "pino-pretty": "^7.3.0",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.2",
     "uuid": "8.3.2",
     "yamljs": "0.3.0"
   },
@@ -78,7 +82,7 @@
     "prettier": "2.4.1",
     "supertest": "6.1.6",
     "ts-node-dev": "^1.1.8",
-    "typedoc": "^0.22.10",
-    "typescript": "^4.5.2"
+    "tsc-watch": "^4.6.0",
+    "typedoc": "^0.22.10"
   }
 }
