+++ src/resources/users/user.memory.repository.ts
@@ -1,5 +1,5 @@
 import { IUser } from '../../common/interfaces';
-import { users } from '../../dataBase/users.db';
+import { users } from '../../../dataBase/users.db';
 
 let allUsers: Array<IUser> = [...users];
